move assertList* dataproviders to seperate class
