+++ tests/php/Dev/SapphireTestTest.php
@@ -60,7 +60,7 @@ class SapphireTestTest extends SapphireTest
     }
 
     /**
-     * @dataProvider provideEqualLists
+     * @dataProvider \SilverStripe\Dev\Tests\SapphireTestTest\DataProvider::provideEqualLists
      *
      * @param $matches
      * @param $itemsForList
@@ -93,106 +93,9 @@ class SapphireTestTest extends SapphireTest
         $this->markTestIncomplete();
     }
 
-    public function provideEqualLists()
-    {
-        $oneItemList = [
-            ['FirstName' => 'Ingo', 'Surname' => 'Schommer']
-        ];
-        $twoItemList = [
-            ['FirstName' => 'Ingo', 'Surname' => 'Schommer'],
-            ['FirstName' => 'Sam', 'Surname' => 'Minnee']
-        ];
-
-        return [
-            [ //empty list
-                [],
-                []
-            ],
-            [
-                [ //one param
-                    ['FirstName' => 'Ingo']
-                ]
-                , $oneItemList
-            ],
-            [
-                [ //two params
-                    ['FirstName' => 'Ingo', 'Surname' => 'Schommer']
-                ],
-                $oneItemList
-            ],
-            [ //only one param
-                [
-                    ['FirstName' => 'Ingo'],
-                    ['FirstName' => 'Sam']
-                ]
-                , $twoItemList
-            ],
-            [
-                [ //two params
-                    ['FirstName' => 'Ingo', 'Surname' => 'Schommer'],
-                    ['FirstName' => 'Sam', 'Surname' => 'Minnee']
-                ],
-                $twoItemList
-            ],
-            [
-                [ //mixed
-                    ['FirstName' => 'Ingo', 'Surname' => 'Schommer'],
-                    ['FirstName' => 'Sam']
-                ],
-                $twoItemList
-            ],
-        ];
-    }
-    public function provideNonEqualLists()
-    {
-        $oneItemList = [
-            ['FirstName' => 'Ingo', 'Surname' => 'Schommer']
-        ];
-        $twoItemList = [
-            ['FirstName' => 'Ingo', 'Surname' => 'Schommer'],
-            ['FirstName' => 'Sam', 'Surname' => 'Minnee']
-        ];
-
-        return [
-            [ //empty list
-                [
-                    ['FirstName' => 'Ingo']
-                ],
-                []
-            ],
-            [
-                [ //one item expected
-                    ['FirstName' => 'Ingo']
-                ]
-                , $twoItemList
-            ],
-            [ //one item with wrong param
-                [
-                    ['FirstName' => 'IngoXX'],
-                    ['FirstName' => 'Sam']
-                ]
-                , $twoItemList
-            ],
-            [
-                [ //two params wrong
-                    ['FirstName' => 'IngoXXX', 'Surname' => 'Schommer'],
-                    ['FirstName' => 'Sam', 'Surname' => 'MinneeXXX']
-                ],
-                $twoItemList
-            ],
-            [
-                [ //mixed
-                    ['FirstName' => 'Daniel', 'Surname' => 'Foo'],
-                    ['FirstName' => 'Dan']
-                ],
-                $twoItemList
-            ],
-        ];
-    }
-
 
     /**
-     * @dataProvider provideEqualLists
+     * @dataProvider \SilverStripe\Dev\Tests\SapphireTestTest\DataProvider::provideEqualLists
      * @testdox Has assertion assertListEquals
      *
      * @param $matches
@@ -210,7 +113,7 @@ class SapphireTestTest extends SapphireTest
     }
 
     /**
-     * @dataProvider provideNonEqualLists
+     * @dataProvider \SilverStripe\Dev\Tests\SapphireTestTest\DataProvider::provideNonEqualLists
      * @testdox assertion assertListEquals fails on non equal Lists
      *
      * @param $matches
